pyproject.toml
[project]
name = "remood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==23.2.1",
    "cryptography==42.0.2",
    "fastapi==0.109.2",
    "jinja2==3.1.3",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.6.1",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "sqlalchemy==2.0.27",
    "uvicorn==0.27.1",
]
